Base Directory
This is the starting point for the website. At the moment it is fully
functional so be careful when committing code as rollbacks and bug
fixing take ages.
At the moment the index page is a php file which means that if you want
to run/develop this on your own machine you will need to setup a
localhost server. While the php functionality isn’t being used at the
moment, if we want any cookie/session functionality we will need it.
Enjoy!